Move blueprint persistance out of the StoreHub
